+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/core/api/SubmittedResource.kt
@@ -36,10 +36,10 @@ data class SubmittedResource<T : ResourceSpec>(
   @Description("The specification of the resource")
   @JsonTypeInfo(use = Id.NAME, include = As.EXTERNAL_PROPERTY, property = "kind")
   val spec: T
-) {
-  val id: String
-    get() = generateId(kind, spec)
-}
+)
+
+val SubmittedResource<*>.id: String
+  get() = generateId(kind, spec)
 
 fun <T : ResourceSpec> SubmittedResource<T>.normalize(): Resource<T> =
   Resource(
