+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/persistence/ArtifactRepository.kt
@@ -10,6 +10,7 @@ import com.netflix.spinnaker.keel.core.api.EnvironmentArtifactVetoes
 import com.netflix.spinnaker.keel.core.api.EnvironmentSummary
 import com.netflix.spinnaker.keel.core.api.PinnedEnvironment
 import com.netflix.spinnaker.keel.core.comparator
+import com.netflix.spinnaker.kork.exceptions.UserException
 import java.time.Duration
 
 interface ArtifactRepository : PeriodicallyCheckedRepository<DeliveryArtifact> {
@@ -262,4 +263,22 @@ interface ArtifactRepository : PeriodicallyCheckedRepository<DeliveryArtifact> {
 
     return pending.filter { isOlder(artifact, it, currentVersion) }
   }
+}
+
+class NoSuchArtifactException(name: String, type: ArtifactType) :
+  NoSuchEntityException("No $type artifact named $name is registered") {
+  constructor(artifact: DeliveryArtifact) : this(artifact.name, artifact.type)
+}
+
+class ArtifactReferenceNotFoundException(deliveryConfig: String, reference: String, type: ArtifactType) :
+  NoSuchEntityException("No $type artifact with reference $reference in delivery config $deliveryConfig is registered")
+
+class ArtifactNotFoundException(name: String, type: ArtifactType, reference: String, deliveryConfig: String?) :
+  NoSuchEntityException("No $type artifact named $name with reference $reference in delivery config $deliveryConfig is registered") {
+  constructor(artifact: DeliveryArtifact) : this(artifact.name, artifact.type, artifact.reference, artifact.deliveryConfigName)
+}
+
+class ArtifactAlreadyRegistered(name: String, type: ArtifactType) :
+  UserException("The $type artifact $name is already registered") {
+  constructor(artifact: DeliveryArtifact) : this(artifact.name, artifact.type)
 }
