+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/persistence/NoSuchEntityException.kt
@@ -1,0 +1,6 @@
+package com.netflix.spinnaker.keel.persistence
+
+import com.netflix.spinnaker.kork.exceptions.SystemException
+
+abstract class NoSuchEntityException(override val message: String?) :
+  SystemException(message)
